requirements.txt
flask
flask-login
adafruit-circuitpython-dht
RPi.GPIO
